Move lsxfel module into cli folder
